requirements.txt
annotated-types==0.7.0
fastapi==0.115.12
certifi==2026.2.25
click==8.1.8
cffi==2.0.0
anyio==4.9.0
charset-normalizer==3.4.4
cryptography==46.0.5
idna==3.11
iniconfig==2.3.0
joblib==1.5.3
linkify-it-py==2.0.3
markdown-it-py==4.0.0
mdit-py-plugins==0.5.0
mdurl==0.1.2
numpy==2.4.2
packaging==26.0
platformdirs==4.9.2
pluggy==1.6.0
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pytest==9.0.2
python-dotenv==1.2.1
requests==2.32.5
rich==14.3.3
scikit-learn==1.8.0
scipy==1.17.1
textual==8.0.0
textual-plotext==1.0.1
threadpoolctl==3.6.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uc-micro-py==1.0.3
urllib3==2.6.3
uvicorn[standard]==0.34.3
websockets==16.0
